+++ python/tvm/autotvm/task/__init__.py
@@ -11,5 +11,5 @@ from .space import ConfigSpace, ConfigEntity
 from .code_hash import attach_code_hash, attach_code_hash_to_arg
 from .dispatcher import DispatchContext, ApplyConfig, dispatcher
 
-from .nnvm_integration import register_topi_compute, register_topi_schedule,\
-        extract_from_graph
+from .topi_integration import register_topi_compute, register_topi_schedule
+from .nnvm_integration import extract_from_graph
