+++ python/tvm/autotvm/tuner/graph_tuning.py
@@ -58,7 +58,7 @@ def tune_tasks(tasks,
 
         # create tuner
         if tuner == 'xgb' or tuner == 'xgb-rank':
-            tuner_obj = XGBTuner(tsk, loss_type='rank', verbose=0)
+            tuner_obj = XGBTuner(tsk, loss_type='rank')
         elif tuner == 'ga':
             tuner_obj = GATuner(tsk, pop_size=50)
         elif tuner == 'random':
